セッションチェック処理をモジュール化した
