+++ taskManager/app/helpers/login_helper.rb
@@ -1,2 +1,5 @@
 module LoginHelper
+  def logged_in?
+    check_session_data
+  end
 end
